requirements.txt
transformers
torch
pdfplumber
python-docx
reportlab